requirements.txt
fontawesomefree==6.4.0
django-ckeditor
channels